Finished Removing React Code from combineNodes
